Change: Move AnimatedObject to API
