+++ source/OpenBVE/Game/ObjectManager/AnimatedObjects/AnimatedWorldObject.cs
@@ -6,43 +6,6 @@ namespace OpenBve
 	/// <summary>The ObjectManager is the root class containing functions to load and manage objects within the simulation world</summary>
 	public static partial class ObjectManager
 	{
-		private class AnimatedWorldObject : WorldObject
-		{
-			/// <summary>The actual animated object</summary>
-			internal AnimatedObject Object;
-			/// <summary>The signalling section the object refers to (Only relevant for objects placed using Track.Sig</summary>
-			internal int SectionIndex;
-
-			public override void Update(AbstractTrain NearestTrain, double TimeElapsed, bool ForceUpdate, bool Visible)
-			{
-				if (Visible | ForceUpdate)
-				{
-					if (Object.SecondsSinceLastUpdate >= Object.RefreshRate | ForceUpdate)
-					{
-						double timeDelta = Object.SecondsSinceLastUpdate + TimeElapsed;
-						Object.SecondsSinceLastUpdate = 0.0;
-						Object.Update(false, NearestTrain, NearestTrain == null ? 0 : NearestTrain.DriverCar, SectionIndex, TrackPosition, Position, Direction, Up, Side, true, true, timeDelta, true);
-					}
-					else
-					{
-						Object.SecondsSinceLastUpdate += TimeElapsed;
-					}
-					if (!base.Visible)
-					{
-						Renderer.ShowObject(Object.internalObject, ObjectType.Dynamic);
-						base.Visible = true;
-					}
-				}
-				else
-				{
-					Object.SecondsSinceLastUpdate += TimeElapsed;
-					if (base.Visible)
-					{
-						Renderer.HideObject(ref Object.internalObject);
-						base.Visible = false;
-					}
-				}
-			}
-		}
+		
 	}
 }
