+++ source/OpenBVE/Game/ObjectManager/ObjectManager.cs
@@ -63,7 +63,7 @@ namespace OpenBve
 					}
 				}
 				const double extraRadius = 10.0;
-				double z = AnimatedWorldObjects[i].Object.TranslateZFunction == null ? 0.0 : Object.TranslateZFunction.LastResult;
+				double z = AnimatedWorldObjects[i].Object.TranslateZFunction == null ? 0.0 : AnimatedWorldObjects[i].Object.TranslateZFunction.LastResult;
 				double pa = AnimatedWorldObjects[i].TrackPosition + z - AnimatedWorldObjects[i].Radius - extraRadius;
 				double pb = AnimatedWorldObjects[i].TrackPosition + z + AnimatedWorldObjects[i].Radius + extraRadius;
 				double ta = World.CameraTrackFollower.TrackPosition + Camera.Alignment.Position.Z - Backgrounds.BackgroundImageDistance - Camera.ExtraViewingDistance;
