+++ source/OpenBVE/OpenBve.csproj
@@ -129,7 +129,6 @@
       <DependentUpon>MessageManager.cs</DependentUpon>
     </Compile>
     <Compile Include="Game\Messages.cs" />
-    <Compile Include="Game\ObjectManager\AnimatedObjects\AnimatedObject.cs" />
     <Compile Include="Game\ObjectManager\AnimatedObjects\AnimatedObjectCollection.cs" />
     <Compile Include="Game\ObjectManager\AnimatedObjects\AnimatedWorldObject.cs" />
     <Compile Include="Game\ObjectManager\AnimatedObjects\AnimatedWorldObject.StateSound.cs" />
