+++ source/OpenBveApi/OpenBveApi.csproj
@@ -99,6 +99,7 @@
     <Compile Include="Objects\ObjectInterface.cs" />
     <Compile Include="Objects\ObjectTypes\AnimatedObject.AnimatedObjectState.cs" />
     <Compile Include="Objects\ObjectTypes\AnimatedObject.Base.cs" />
+    <Compile Include="Objects\ObjectTypes\AnimatedWorldObject.cs" />
     <Compile Include="Objects\ObjectTypes\StaticObject.cs" />
     <Compile Include="Objects\ObjectTypes\TrackFollowingObject.cs" />
     <Compile Include="Objects\ObjectTypes\UnifiedObject.cs" />
